lead_qual_crew/crewai_plus_lead_scoring/requirements.txt
crewai>=0.11.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # For PostgreSQL support
python-dotenv>=1.0.0    # For environment variable management 
crewai-tools>=0.38.1
openai>=1.13.3
fastapi
uvicorn[standard]
PyYAML
crewai-tools>=0.1.0     # Add crewai-tools dependency 
